sensorPI/requirements.txt
schedule
paho.mqtt.client
json
bs4.BeautifulSoup
geopy
pymongo
requests
configparser
flask
context
influxdb
threading
pymongo